Create NPF devices as filter modules are attached
Previous behavior was to go through the Registry for adapter names and
create NPF devices for each one, then correlate the name back to the
appropriate filter module that had been attached to that adapter.
Instead, we now create the NPF device when the filter module is
attached, meaning we use the same source of names for both NPF devices
and filter modules. This saves a trip to the Registry and ensures a
1-to-1 relationship between NPF devices and filter modules.